+++ Frends.MicrosoftSQL.ExecuteQuery/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.1] - 2023-01-27
+### Changed
+- Naming: Result.QueryResult to Result.Data.
+
 ## [1.0.0] - 2023-01-18
 ### Added
 - Initial implementation
