+++ Frends.MicrosoftSQL.ExecuteQuery/Frends.MicrosoftSQL.ExecuteQuery.Tests/docker-compose.yml
@@ -1,0 +1,13 @@
+version: '3'
+
+services:
+  sql1:
+    container_name: sql1
+    image: mcr.microsoft.com/mssql/server:2019-CU18-ubuntu-20.04    
+    hostname:
+        sql1 
+    environment:
+      MSSQL_SA_PASSWORD: "Salakala123!"
+      ACCEPT_EULA: "Y"
+    ports:
+      - "1433:1433"
