+++ Frends.MicrosoftSQL.ExecuteQuery/Frends.MicrosoftSQL.ExecuteQuery/ExecuteQuery.cs
@@ -33,7 +33,7 @@ public class MicrosoftSQL
     /// <param name="input">Input parameters</param>
     /// <param name="options">Optional parameters</param>
     /// <param name="cancellationToken">Token generated by Frends to stop this Task.</param>
-    /// <returns>Object { bool Success, int RecordsAffected, string ErrorMessage, JToken QueryResult }</returns>
+    /// <returns>Object { bool Success, int RecordsAffected, string ErrorMessage, JToken Data }</returns>
     public static async Task<Result> ExecuteQuery([PropertyTab] Input input, [PropertyTab] Options options, CancellationToken cancellationToken)
     {
         Result result;
@@ -142,37 +142,30 @@ public class MicrosoftSQL
 
             if (command.Transaction is null)
             {
-                var eMsg = $"ExecuteHandler exception: 'Options.SqlTransactionIsolationLevel = None', so there was no transaction rollback. {ex}.";
                 if (options.ThrowErrorOnFailure)
-                    throw new Exception(eMsg);
+                    throw new Exception("ExecuteHandler exception: 'Options.SqlTransactionIsolationLevel = None', so there was no transaction rollback.", ex);
                 else
-                    return new Result(false, 0, eMsg, null);
+                    return new Result(false, 0, $"ExecuteHandler exception: 'Options.SqlTransactionIsolationLevel = None', so there was no transaction rollback. {ex}", null);
             }
             else
             {
-                if (dataReader != null && !dataReader.IsClosed)
-                    await dataReader.CloseAsync();
-
                 try
                 {
                     await command.Transaction.RollbackAsync(cancellationToken);
                 }
                 catch (Exception rollbackEx)
                 {
-                    var rbEMsg = $"ExecuteHandler exception: An exception occurred on transaction rollback. Rollback exception: {rollbackEx}. ||  Original exception leading to rollback: {ex}";
                     if (options.ThrowErrorOnFailure)
-                        throw new Exception(rbEMsg);
+                        throw new Exception("ExecuteHandler exception: An exception occurred on transaction rollback.", rollbackEx);
                     else
-                        return new Result(false, 0, rbEMsg, null);
+                        return new Result(false, 0, $"ExecuteHandler exception: An exception occurred on transaction rollback. Rollback exception: {rollbackEx}. ||  Exception leading to rollback: {ex}", null);
                 }
 
-                var eMsg = $"ExecuteHandler exception: (If required) transaction rollback completed without exception. {ex}.";
                 if (options.ThrowErrorOnFailure)
-                    throw new Exception(eMsg);
+                    throw new Exception("ExecuteHandler exception: (If required) transaction rollback completed without exception.", ex);
                 else
-                    return new Result(false, 0, eMsg, null);
+                    return new Result(false, 0, $"ExecuteHandler exception: (If required) transaction rollback completed without exception. {ex}.", null);
             }
-
         }
     }
